Move top-level loop up to Application level and get rid of Execution project.
