+++ Symbolica.sln
@@ -22,10 +22,6 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Deserialization", "src\Dese
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Deserialization.Tests", "tests\Deserialization.Tests\Deserialization.Tests.csproj", "{57B13108-9D2F-49B9-9067-553D790E5835}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Execution", "src\Execution\Execution.csproj", "{C508B3DD-64F8-45E9-B2F8-D49EB6595150}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Execution.Tests", "tests\Execution.Tests\Execution.Tests.csproj", "{927BBB85-0FB2-4955-90A6-E40977C92C2B}"
-EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Computation", "src\Computation\Computation.csproj", "{1877D419-C4BF-4C76-930C-009AF3FF89F2}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Computation.Tests", "tests\Computation.Tests\Computation.Tests.csproj", "{A21B3F62-B861-4885-B5B1-F2D156F8D6C2}"
@@ -49,8 +45,6 @@ Global
 		{D306C291-2B73-4F26-A4C7-C8B4C13EC1BA} = {D31C471B-BC9B-46DD-901A-400AFA1656CC}
 		{857D7336-99FC-4D2E-A3ED-4B9F5BA3C140} = {11AB1280-FD82-4CEA-B6EA-91BE7805A4E8}
 		{57B13108-9D2F-49B9-9067-553D790E5835} = {D31C471B-BC9B-46DD-901A-400AFA1656CC}
-		{C508B3DD-64F8-45E9-B2F8-D49EB6595150} = {11AB1280-FD82-4CEA-B6EA-91BE7805A4E8}
-		{927BBB85-0FB2-4955-90A6-E40977C92C2B} = {D31C471B-BC9B-46DD-901A-400AFA1656CC}
 		{1877D419-C4BF-4C76-930C-009AF3FF89F2} = {11AB1280-FD82-4CEA-B6EA-91BE7805A4E8}
 		{A21B3F62-B861-4885-B5B1-F2D156F8D6C2} = {D31C471B-BC9B-46DD-901A-400AFA1656CC}
 		{9DBA5223-8FB2-4240-89E7-EC4CB382703C} = {11AB1280-FD82-4CEA-B6EA-91BE7805A4E8}
@@ -93,14 +87,6 @@ Global
 		{57B13108-9D2F-49B9-9067-553D790E5835}.Debug|x64.Build.0 = Debug|x64
 		{57B13108-9D2F-49B9-9067-553D790E5835}.Release|x64.ActiveCfg = Release|x64
 		{57B13108-9D2F-49B9-9067-553D790E5835}.Release|x64.Build.0 = Release|x64
-		{C508B3DD-64F8-45E9-B2F8-D49EB6595150}.Debug|x64.ActiveCfg = Debug|x64
-		{C508B3DD-64F8-45E9-B2F8-D49EB6595150}.Debug|x64.Build.0 = Debug|x64
-		{C508B3DD-64F8-45E9-B2F8-D49EB6595150}.Release|x64.ActiveCfg = Release|x64
-		{C508B3DD-64F8-45E9-B2F8-D49EB6595150}.Release|x64.Build.0 = Release|x64
-		{927BBB85-0FB2-4955-90A6-E40977C92C2B}.Debug|x64.ActiveCfg = Debug|x64
-		{927BBB85-0FB2-4955-90A6-E40977C92C2B}.Debug|x64.Build.0 = Debug|x64
-		{927BBB85-0FB2-4955-90A6-E40977C92C2B}.Release|x64.ActiveCfg = Release|x64
-		{927BBB85-0FB2-4955-90A6-E40977C92C2B}.Release|x64.Build.0 = Release|x64
 		{1877D419-C4BF-4C76-930C-009AF3FF89F2}.Debug|x64.ActiveCfg = Debug|x64
 		{1877D419-C4BF-4C76-930C-009AF3FF89F2}.Debug|x64.Build.0 = Debug|x64
 		{1877D419-C4BF-4C76-930C-009AF3FF89F2}.Release|x64.ActiveCfg = Release|x64
