+++ src/Application/Implementation/IAwaitableProgramPool.cs
@@ -1,0 +1,10 @@
+﻿using System.Threading.Tasks;
+using Symbolica.Implementation;
+
+namespace Symbolica.Application.Implementation
+{
+    internal interface IAwaitableProgramPool : IProgramPool
+    {
+        Task Wait();
+    }
+}
