+++ src/Application/Program.cs
@@ -8,7 +8,6 @@ using Symbolica.Application.Computation;
 using Symbolica.Application.Implementation;
 using Symbolica.Computation;
 using Symbolica.Deserialization;
-using Symbolica.Execution;
 using Symbolica.Implementation;
 using Symbolica.Implementation.System;
