--- src/Execution/Execution.csproj
@@ -1,13 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <AssemblyName>Symbolica.Execution</AssemblyName>
-    <RootNamespace>Symbolica.Execution</RootNamespace>
-    <TargetFramework>net5.0</TargetFramework>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\Abstraction\Abstraction.csproj" />
-  </ItemGroup>
-
-</Project>
