--- src/Execution/IAwaitableProgramPool.cs
@@ -1,9 +1,0 @@
-﻿using System.Threading.Tasks;
-
-namespace Symbolica.Execution
-{
-    public interface IAwaitableProgramPool : IProgramPool
-    {
-        Task Wait();
-    }
-}
