+++ src/Implementation/Program.cs
@@ -1,5 +1,4 @@
 ﻿using System;
-using Symbolica.Execution;
 
 namespace Symbolica.Implementation
 {
