+++ src/Implementation/ProgramFactory.cs
@@ -2,7 +2,6 @@
 using System.Linq;
 using Symbolica.Abstraction;
 using Symbolica.Collection;
-using Symbolica.Execution;
 using Symbolica.Expression;
 using Symbolica.Implementation.Memory;
 using Symbolica.Implementation.Stack;
