--- tests/Execution.Tests/Execution.Tests.csproj
@@ -1,20 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <AssemblyName>Symbolica.Execution.Tests</AssemblyName>
-    <RootNamespace>Symbolica.Execution</RootNamespace>
-    <TargetFramework>net5.0</TargetFramework>
-   </PropertyGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\..\src\Execution\Execution.csproj" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <PackageReference Include="FluentAssertions" Version="5.10.3" />
-    <PackageReference Include="Microsoft.NET.Test.Sdk" Version="16.8.3" />
-    <PackageReference Include="xunit" Version="2.4.1" />
-    <PackageReference Include="xunit.runner.visualstudio" Version="2.4.3" />
-  </ItemGroup>
-
-</Project>
